fix(ui): address review feedback — remove dead code, deduplicate utils and advisories (TC-4523)
- Remove unused RemediationDetails.tsx component
- Remove stale advisory field from RemediationInfo interface
- Extract compareVersions and formatRange to shared utils/version.ts
- Deduplicate advisories by ID in mergeRemediation
- Replace array index key with composite key in AdvisoryRemediations
